move SizeParamGatherer to pytato.transform
